requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
streamlit
httpx
pydantic
python-multipart